Add files via upload
Audacity file (aup3) where there is generated white noise and this was played through a speaker into an tube with flowing gas and recorded with a microphone on the other side of the tube.  There are 5 audio tracks associated with this audacity file, one for the white noise and 4 for each of the gases: Air, Nitrogen, Argon and an 'Unknown' gas (Hint: is a common refrigerant gas).